use safe calls of external tools
Make calls of external tools in mlxacl and mlxrtr safe. Commands and
their options were called as a single string without additional
processing of arguments. It has not caused problems because all of that
is shell-safe. Anyway using explicit arguments with proper quoting is
safer and we do not need to remember keeping the arguments safe.