make mac addresses on some interfaces persistent
MAC address of interfaces like bond, vlan or switch can be changed
automatically when some interface changes occur: a port added to the bond,
an interface added to the switch, etc. For example port split can
introduce new interfaces to the system, which are added to the switch,
and that can cause change of MACs on all vlan interfaces, which is
undesirable on a running system.
This change verifies wether MACs on such interfaces are set manually,
and if not - replaces the MAC on the interface with the same value
so that is should remain permanent during the further changes.